PROJECT/requirements.txt
beautifulsoup4==4.8.1
bs4==0.0.1
Click==7.0
Flask==1.1.1
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
mysql-connector==2.2.9
soupsieve==1.9.5
Werkzeug==0.16.0